feat: implement cookie store functions for token management and add refresh token handling in apiJson
