+++ web/app/lib/apiClient.ts
@@ -1,30 +1,11 @@
 import { API_URL } from "~/config/ApiConfig";
+import { getAccessToken, getRefreshToken, saveCookieStoreValue } from "./GetCookie";
 
 type ApiFetchOptions = Omit<RequestInit, "body" | "headers"> & {
   body?: unknown;
   headers?: HeadersInit;
   auth?: boolean;
 };
-
-async function readCookieStoreValue(key: string): Promise<string | null> {
-  if (typeof window === "undefined") return null;
-  const anyWindow = window as any;
-  if (!anyWindow.cookieStore?.get) return null;
-  try {
-    const cookie = await anyWindow.cookieStore.get(key);
-    return cookie?.value ?? null;
-  } catch {
-    return null;
-  }
-}
-
-export async function getAccessToken(): Promise<string | null> {
-  const fromCookieStore = await readCookieStoreValue("token");
-  if (fromCookieStore) return fromCookieStore;
-
-  if (typeof window === "undefined") return null;
-  return localStorage.getItem("token");
-}
 
 function buildUrl(path: string) {
   if (!API_URL) return path;
@@ -80,6 +61,23 @@ export async function apiJson<T>(path: string, options: ApiFetchOptions = {}): P
     headers,
     body: options.body !== undefined ? JSON.stringify(options.body) : undefined,
   });
+
+  if(response.status === 401) {
+    const refresh = await getRefreshToken();
+    if (!refresh) {
+      throw new Error("No hay token de sesión; inicia sesión nuevamente");
+    }
+    const refreshResponse = await fetch(buildUrl("/auth/refresh-token"), {
+      method: "POST",
+      body: JSON.stringify({ refresh_token: refresh }),
+      headers: { "Content-Type": "application/json" },
+    }).then(res => res.json());
+
+    if (refreshResponse?.session?.access_token) {
+      await saveCookieStoreValue("token", refreshResponse.session.access_token);
+      return apiJson(path, options);
+    }
+  }
 
   const contentType = response.headers.get("content-type") || "";
   const isJson = contentType.includes("application/json");
